heroku app/requirements.txt
Flask==0.11.1
Flask-PyMongo==0.4.1
itsdangerous==0.24
gunicorn==19.6.0
Jinja2==2.8
MarkupSafe==0.23
nltk==3.2.1
pymongo==3.3.1
python-highcharts==0.2.0
Werkzeug==0.11.11
